thursday meeting changes
