+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CompetitionTeleOp.java
@@ -7,6 +7,7 @@ import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import org.firstinspires.ftc.teamcode.Shared.Drive2;
+import org.firstinspires.ftc.teamcode.robot.CompetitionRobot;
 import org.firstinspires.ftc.teamcode.robot.hardware.Arm.HubPosition;
 import org.firstinspires.ftc.teamcode.util.ExMath;
 import org.firstinspires.ftc.teamcode.robot.Robot;
@@ -14,7 +15,7 @@ import org.firstinspires.ftc.teamcode.robot.Robot;
 @TeleOp(name = "Competition TeleOp", group = "Competition")
 public class CompetitionTeleOp extends LinearOpMode {
     //private final String TAG = getClass().getName();
-    Robot robot;
+    CompetitionRobot robot;
     DcMotor Arm = null;
     Servo Wrist = null;
     CRServo Collector = null;
@@ -52,7 +53,7 @@ public class CompetitionTeleOp extends LinearOpMode {
         rightCarousel = hardwareMap.get(CRServo.class, "rightCarousel");
         leftCarousel = hardwareMap.get(CRServo.class, "leftCarousel");
         cap = hardwareMap.get(Servo.class, "cap");
-        robot = new Robot(hardwareMap);
+        robot = new CompetitionRobot(hardwareMap);
         robot.initHardware();
         drive = new Drive2(robot, this);
         cap.setPosition(.5);
