+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SpeedChangeTest.java
@@ -5,6 +5,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.Shared.Drive2;
+import org.firstinspires.ftc.teamcode.robot.CompetitionRobot;
 import org.firstinspires.ftc.teamcode.robot.Robot;
 
 @Disabled
@@ -14,7 +15,7 @@ public class SpeedChangeTest extends LinearOpMode {
     public void runOpMode() throws InterruptedException {
 
         // Create the robot from the hardware map
-        Robot robot = new Robot(hardwareMap);
+        CompetitionRobot robot = new CompetitionRobot(hardwareMap);
         // Initialize the hardware
         robot.initDrives();
         robot.initIMU();
