+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/TopLeft.java
@@ -1,12 +1,24 @@
 package org.firstinspires.ftc.teamcode.autonomous;
 
 import com.qualcomm.hardware.rev.RevBlinkinLedDriver;
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
+import org.firstinspires.ftc.teamcode.Shared.Drive2;
+import org.firstinspires.ftc.teamcode.robot.TestRobot;
+
+@Autonomous
 public class TopLeft extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
-        RevBlinkinLedDriver lights = hardwareMap.get(RevBlinkinLedDriver.class, "LED");
-        lights.setPattern(RevBlinkinLedDriver.BlinkinPattern.RED);
+        TestRobot robot = new TestRobot(hardwareMap);
+
+        robot.initHardware();
+
+        Drive2 drive = new Drive2(robot, this);
+
+        drive.navigationMonitorTicks(10, 10, 10, 10);
+
+
     }
 }
