+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/CompetitionRobot.java
@@ -1,28 +1,40 @@
 package org.firstinspires.ftc.teamcode.robot;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
+import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
+import com.qualcomm.robotcore.hardware.Servo;
 
+import org.firstinspires.ftc.teamcode.robot.hardware.Arm;
 import org.firstinspires.ftc.teamcode.robot.hardware.Drive;
+import org.firstinspires.ftc.teamcode.util.HardwareNotFoundException;
+import org.firstinspires.ftc.teamcode.util.Logger;
 import org.firstinspires.ftc.teamcode.util.Timeout;
 
 import java.util.EnumMap;
 
-public class testRobot {
+public class CompetitionRobot extends Robot {
 
     // Hardware definitions
     static final private class Hardware {
         static final double revCounts = 28;
-        static final double gearReduction = 40;
-        static final double wheelDiameter = 3;
+        static final double gearReduction = 40;//40 before
+        static final double wheelDiameter = 6;//3 for test robot
 
-        static final String frontLeftMotorName = "FL";//left_front_drive
-        static final String frontRightMotorName = "FR";//right_front_drive
-        static final String backLeftMotorName = "BL";//left_back_drive
-        static final String backRightMotorName = "BR";//right_back_drive
+        //Competition Robot
+//        static final String frontLeftMotorName = "FL"; // CHM2
+//        static final String frontRightMotorName = "FR"; // CHM3
+//        static final String backLeftMotorName = "BL"; // CHM0
+//        static final String backRightMotorName = "BR"; // CHM1
+//        static final String imuName = "imu";
+
+        static final String frontLeftMotorName = "FL"; // CHM2 //left_front_drive
+        static final String frontRightMotorName = "FR"; // CHM3 //right_front_drive
+        static final String backLeftMotorName = "BL"; // CHM0 //left_back_drive
+        static final String backRightMotorName = "BR"; // CHM1 //right_back_drive
         static final String imuName = "imu";
     }
 
@@ -39,13 +51,28 @@ public class testRobot {
 
     public BNO055IMU imu;
 
-    public testRobot(HardwareMap hardwareMap) {
+    public Arm arm;
+
+    private final Logger logger;
+
+    public CompetitionRobot(HardwareMap hardwareMap, Logger logger) {
         this.hardwareMap = hardwareMap;
+        this.logger = logger;
     }
 
-    public void initHardware() {
-        initDrives();
-        initIMU();
+    public CompetitionRobot(HardwareMap hardwareMap) {
+        this.hardwareMap = hardwareMap;
+        this.logger = new Logger();
+    }
+
+    public void initHardware() throws IllegalArgumentException {
+        try {
+            initDrives();
+            initArm();
+            initIMU();
+        } catch (IllegalArgumentException e) {
+            throw new HardwareNotFoundException(e);
+        }
     }
 
     public void initIMU() {
@@ -73,19 +100,28 @@ public class testRobot {
         }
     }
 
+    public void initArm() {
+        this.arm = new Arm(
+                this.hardwareMap.get(DcMotor.class, "Arm"), // EXM0
+                this.hardwareMap.get(Servo.class, "pivotCollector"), // CHS3
+                this.hardwareMap.get(CRServo.class, "spinCollector") // CHS2
+        );
+    }
+
+
     public void initDrives() {
         this.Drives.put(DrivePos.FRONT_LEFT,
                 new Drive(this.hardwareMap.get(DcMotorEx.class, Hardware.frontLeftMotorName),
-                        Hardware.revCounts, Hardware.gearReduction, Hardware.wheelDiameter));
+                        Hardware.revCounts,Hardware.gearReduction,Hardware.wheelDiameter));
         this.Drives.put(DrivePos.FRONT_RIGHT,
                 new Drive(this.hardwareMap.get(DcMotorEx.class, Hardware.frontRightMotorName),
-                        Hardware.revCounts, Hardware.gearReduction, Hardware.wheelDiameter));
+                        Hardware.revCounts,Hardware.gearReduction,Hardware.wheelDiameter));
         this.Drives.put(DrivePos.BACK_LEFT,
                 new Drive(this.hardwareMap.get(DcMotorEx.class, Hardware.backLeftMotorName),
-                        Hardware.revCounts, Hardware.gearReduction, Hardware.wheelDiameter));
+                        Hardware.revCounts,Hardware.gearReduction,Hardware.wheelDiameter));
         this.Drives.put(DrivePos.BACK_RIGHT,
                 new Drive(this.hardwareMap.get(DcMotorEx.class, Hardware.backRightMotorName),
-                        Hardware.revCounts, Hardware.gearReduction, Hardware.wheelDiameter));
+                        Hardware.revCounts,Hardware.gearReduction,Hardware.wheelDiameter));
 
         for (Drive drive : this.Drives.values())
             drive.run();
